Support `gradle.properties` version file
The `gradle.properties` file allows for a default version while still allowing for the version to be overriden on the command line.  This is a perfect scenario for new-release-version

For example:
Set the property only in the `gradle.properties` file (i.e. remove it from build.gradle). Also make sure the options come before the command (as mentioned above).

gradle.properties contents:

```
version=1.0.12
```

Version can then be overridden on the command line with:

```
gradle -Pversion=1.0.13 publish
```

Based on the StackOverflow answer https://stackoverflow.com/a/40239092/620196